% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typehint.r
\name{typehint}
\alias{typehint}
\title{Package 'typehint'}
\description{
Automatically check the data type, number and size of dimensions, and values of
function arguments with simple type hint comments in the function code.
}
\details{
Type hints are special comments with a leading \code{#|} within a
function body indicating the intended nature of the function's arguments in
terms of data types, dimensions and even permitted values. The actual
parameters with which the function is called can be evaluated against these
type hint comments using the \code{check_types()} function.
}
\section{Author}{

Joachim Zuckarelli, joachim@zuckarelli.de, @jsugarelli
}

\seealso{
Other typehint: 
\code{\link{check_types}()},
\code{\link{show_typehints}()}
}
\concept{typehint}
